requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfminer.six==20221105
python-docx==0.8.11
scikit-learn==1.3.0
numpy==1.24.3
Werkzeug==2.3.7